requirements.txt
#
librosa==0.8.1
pandas==1.2.5
numpy==1.19
scipy==1.6.1
scikit-learn==0.24.2
matplotlib==3.6.3
seaborn==0.12.2
tqdm==4.62.1
pyyaml==6.0
HyperPyYAML==1.2.1
joblib==1.3.1
#
transformers==4.30.2
datasets==2.13.1
accelerate>=0.20.1
evaluate==0.4.0
ipywidgets==8.0.7
#
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
